python-etl/requirements.txt
psycopg2-binary
minio
pandas
pyarrow
sqlalchemy
flask
python-multipart
python-docx
PyMuPDF
pillow
pytesseract
chardet
pdfplumber
python-docx
pypdf
python-pptx
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt
flask_jwt_extended
python-magic
exifread  